+++ src/com/mideas/rpg/v2/game/item/potion/Potion.java
@@ -21,11 +21,11 @@ public class Potion extends Item {
 	private final static String empty = "";
 	
 	public Potion(int id) {
-		super(id, empty, ItemType.POTION, empty, 1, 0, 1, 1);
+		super(id, empty, ItemType.POTION, empty, (byte)1, 0, 1, 1);
 	}
 
 	public Potion(Potion potion) {
-		super(potion.id, potion.sprite_id, ItemType.POTION, potion.name, 1, potion.sellPrice, potion.maxStack, potion.amount);
+		super(potion.id, potion.sprite_id, ItemType.POTION, potion.name, (byte)1, potion.sellPrice, potion.maxStack, potion.amount);
 		this.doHealString = potion.doHealString;
 		this.doManaString = potion.doManaString;
 		this.levelString = potion.levelString;
@@ -35,7 +35,7 @@ public class Potion extends Item {
 	}
 	
 	public Potion(int id, String sprite_id, String name, int level, int doHeal, int doMana, int sellPrice, int amount) {
-		super(id, sprite_id, ItemType.POTION, name, 1, sellPrice, 200, amount);
+		super(id, sprite_id, ItemType.POTION, name, (byte)1, sellPrice, 200, amount);
 		this.doHeal = doHeal;
 		this.doMana = doMana;
 		this.level = level;
